+++ packages/alfa-rules/src/common/predicate/fold-expectation.ts
